+++ client/src/components/AddItemForm/AddItemForm.scss
@@ -1,0 +1,5 @@
+.add-item-form {
+  margin-top: 40px;
+  display: flex;
+  width: 100%;
+}
