+++ client/src/components/TodoList/TodoList.scss
@@ -1,10 +1,4 @@
 .todo-list-container {
-  .add-item-form {
-    margin-top: 40px;
-    display: flex;
-    width: 100%;
-  }
-
   .items {
     margin-top: 40px;
